Make it possible to style the conference registration site
Adds a field to each conference called base_template, which can specify
any template used in place of nav_events.html to inherit from all the
confreg templates. This template in turn can inherit other templates
as necessary.

These other templates can live in a different directory alltogether,
as long as this directory is added to the TEMPLATE_DIRS setting, which
allows us to use a complete framework from a different application on
the same server, provided the naming of these templates is unique
(preferrably using a subdirectory structure to ensure that inheritance
is also unique).