New generation of the invoicing system
This invoice system is designed to be much more than just an invoice generator.
At the core is a more API-based invoice generation, and it has a much easier
to work with interface as well. Let's consider the last version a prototype.

* Supports tracking which invoices are paid
* Supports automatically flagging invoices as paid based on API
  (paypal is the obvious, but it's abstracted to support others as well)
* Supports creation of invoices via internal API, such as conference registrations
* Supports in-place editing and saving of temporary invoices while built up
* Supports connecting an invoice to a community account (not mandatory), making
  it possible for a user to view all their invoices historically
* Automatically generates receipts for all paid invoices. If a community account
  is registered they will be viewable on the web, and as long as an email address
  is registered, it will also be emailed to that.

The idea is that this will act as a hub for *all* payment processing, instead of
having it piece by piece in the different other apps.