Rewrite SSL redirection code
Now uses a decorator instead of manual checks inside each view.

Also add SSL decorators to the new invoice and paypal code to make sure
it's served over SSL.

Finally, define two new settings for SITEBASE and SITEBASE_SSL to override
the website names in a local sandbox install. Makes it possible to for example
test the full paypal workflow, since we need to be able to pass an actual
working URL to paypal in that case.