Prevent deletion/canceling of invoice connected to a system
We don't yet have support for *properly* cancelling invoices generated
from confreg or membership, since those need to make modifications in
the calling systems as well. This needs to be implemented, but for now
just make sure we can't delete it since that would put the system in
a broken state.