Implement bulk payment for conference registrations
Regular users register the normal way, but they never proceed to payment.
Instead, the "bulk payer" will enter their email addresses on a form, confirm
all the payments, and get a single invoice with all of them on it.

The invoicing is integrated with the general invoice system, so once the invoice
is paid, it will automatically flag all the individual registrations as
confirmed as well. (Same goes for cancelling which will unlock them all)