Implement Adyen payment processor module
This squeezes a large number of commits over time, and should at this
point contain a more or less complete Adyen processor. There are some
things around reports that still need to be merged, once we have some
real reports to look at. Notifications are at this point not fully tested,
since we have no place to send them in the testing environment yet.

Cron jobs are not yet merged, that will send summary reports and such
things. Instead, I will verify it manually each day in the beginning,
and then write the scripts :)