Properly process amounts in notifications
Parameters are sent differently in SOAP and POST, and the previous implementation
incorrectly used the SOAP parameter names.

Also, enforce all payemnts in EUR, since that's what the rest of the system expects.

Finally add logging of when we get invalid amounts or currencies. This will fail
at a later stage as well when matching an invoice, but it's good to log every
step of the way.