Don't include the count-of-additional-options annotation in modification requests
This breaks the django ORM when doing admin actions on the queryset. We hack
around this by just not annotating the queryset for POSTs, since they are never
interested in viewing the count anyway - they do updates.

This fixes a crash reported by ads when using the "confirm payments" admin
action.