Process Adyen authorizations on non-primary accounts
This happens for example when receiving a payment from a POS terminal.
We still want to track these transactions in the system, as well as
sending notification emails about them, but we don't want them to try
to find an invoice as they are not automated.

All accounts that aren't the primary will be treated this way, even though
we at this point only have one.