Exclude invoice and reminder fields from membership form
These were included in the form definition but not the webpage, which caused them
to be reset if someone clicked Save on the page. In particular, this means if someone
created an invoice, and then edited their own entry somehow, that invoice would get
detached from the registration. If it was later paid, it would cause an exception in
the invoice processing since it wasn't attached to anything correctly anymore.