Make check for duplicate notifications take merchant accounts into consideration
General notifications carry a unique pspreference, so it doesn't matter for those,
but for report_available type notifications the pspreference is set to the filename,
which will be the same for two reports of the same kind - even if they are for different
merchant accounts. This would cause one of them to be marked a duplicate, and only
the other one being processed at all.