Make invoice and bulkpayments into links instead of dropdowns
Removing either of those manually on a record is likely to cause
inconsistency between the models. Instead, they should be cancelled
using the proper interface. Therefor, make it impossible to remove
them in the admin interface, replacing it with a link (they can of
course still be removed at the SQL level if things are really broken)