Detect paypal donations, and flag them as matched automatically
This makes sure we don't end up with huge amounts of error logs for
each donation before someone manually flags it in the database.

We do the matching simply based on the transaction text. If somebody
else makes a payment with that specific text we'll do a false match,
but this can never happen for actualy payments since they have the
invoice information in the title.