Show the proper buttons and errors when registration type is set to null
This normally never happened, but was exposed when the FOSDEM registration
was opened without any valid registration types. The system would then
first say that payment wasn't required, and if the user returned to the
page later it would let them generate an invoice for a null amount.

Part of this has been broken since day 1 (the part about showing no cost
when the regtype was zeor), part of it was introduced with the integration
of the invoicing system.