Track a lot more Paypal transaction types
Instead of just tracking payments, include donations (which were previously
*sometimes* included but not always, since Paypal labels them differently
for no apparent reason), payments made out, refunds and bank transfers.

These new transaction types just generate accounting records - which are
left open for manual completion and verification - but it removes a number
of manual steps in keeping the numbers up to date. In theory, with this change
everything that happens on the paypal account will get the "paypal side" of
the accounting record automatically filled in.