Make conference reports more filterable
For all fields where we could previously just filter on "must exist",
we'll now dynamically show a textbox that lets you filter (using a
cast insensitive contains operation) on the contents in it. Keep the
checkbox checked and leave the textbox empty for the old behaviour of
just checking if a field is defined.