Add a very simple SQL migration manager
Basically, any app that has a "sqlupdate" subdirectory will be scanned
for files called "00000-<something>.sql", "00001-<something>.sql" etc.
These will be applied in order and the version number from the file will
be stored in a table to make sure that each script is only executed once.
Only the highest number is stored, so only add things to the end. Duplicated
numebers will likely cause confusion, so don't try that...

Scripts are executed *after* syncdb finishes for the app in question, so
references can be used. However, it will run *before* the migration of
the next application on the list, so take care with cross-application
references!

Scripts are run as a whole, so don't make them huge. And of course, all
updates are run in a transaction.