Clean up registration form
This reomves a lot of duplicate code and replaces it with a loop
over an array of fieldsets defined on the form, making it a lot
easier to make changes.

Also update the stylesheets used when an error shows up to make it
a lot easier to spot errors, and possible to render long error
messages without making the formatting really broken.