Fix "negative numbers in balance report" handling
It was exactly 100% reversed... Putting the negative number only in the
sum, and nothing else. Which meant the reports would never match if
anything on the "bottom half" of the balance report was changed during
the year.

This was probably introduced when making it work together with the year
closing functionality, so those have now been fully separated.

Hopefully this fixes all the reports...