Support including open accounting entries in reports
This is optional and crontrolled by a checkbox. Currently only for
balance and results reports.

In doing this, also restructure the get_balance_query function to use
more query generation and less hardcoded strings, as it became completely
unreadable after trying to merge this in. And generally format it for
better readability.