Add partial accounting of Adyen refunds
It's not fully automated. When a refund is sent, an open entry is
created, where the treasurer will have to match it to the original
transaction(s), and make sure the money is taken out of the appropriate
account and object. It's automatically booked against a short-term debt
account.

When the settlement batch completes, we now process the refunded entries
automatically against the rest of that batch.