Add initial support for non-EUR currencies
Put the EUR abbreviation and the € symbol as settings in settings.py
instead, with the default obviously being Euro. We still only support
one currency per installation configured globally, to keep things simple.

The paypal integration requries a separate setting for currency, as
some of that code is not in django (yet).

This is intended in particular to be able to run the confreg system
for non-euro conferences, but the settnigs are made global for the site.