Copy both first and last name to new registrations
This code was not properly updated when the community auth system
started supporting separate fields for first and last name, so it
still assumed a single name and split it. Now that the auth system
updates both first and last name on every login, all details should
be complete, and we can just directly copy them over.