Add functionality to completely override confreg templates
This makes it possible for a conference to completely re-render for example
what a form looks like.

The easiest to do is still to inherit from the base template, and that's what
should mostly be used. But it's now possible to override an individual template
if necessary (for example the registration form), while keeping the base template
for the majority of the pages.