Work around the fact that adyen report notifications are hardcoded to EUR
There is no real currency on report notifications, but Adyen pass along a
hardcoded value of EUR. This rejected the notification when the system was
configured for a non-EUR currency.

With this change, we accept all notifications that are for report availability,
even if they're not in the primary currency, but still reject any other kind
of notifications.