Double check for duplicate paypal transactions
Since our lookup to paypal can take a long time, sometimes a second
notification fires while we're still looking up the first one. To avoid
a key conflict in the db, repeat the check for existing transactions
before writing the new one, and if it's already in there, ignore the
new one. We assume that the result is going to be the same for both
of them.