Clarify even more how to find out what's wrong with the registration form
If a field is incorrectly filled out (e.g. required but missing), it's
highlighted (in yellow/red for example) along with the message. However,
at the top of the form it just said "there's something wrong" which caused
some people not to even check the rest of the form (or miss it even when
they checked). Clarify the text even more to make sure it's obvious where
to look.