Add support for speakers confirming their own presentations
This way we don't have to manually switch talk status twice. With this,
talk status is changed once (submitted -> pending). After this, the
speaker can confirm it him/herself which will move it (pending->approved).

Reserve list is still managed manually, but that should be seldom enough
that it's not a problem.