Add support for automatic notification of accepted/pending sessions
This will not actually send notification emails when the status of
a submission has changed. Instead, the updates will go on a queue,
which will then show up on the admin dashboard and can be sent from
there. The reason for that is to have a safety valve to catch manual
mistakes (as sessions are still changed from submitted->pending
manually).

The email will tell the user what the status is for he new session,
and what this status means, but nothing actually conference specific.
It might be worth adding some sor tof "welcome message" on a per
conference basis in the future.

The notification queue will automatically be hidden once the conference
end date is in the past.