Support splitting a schedule across multiple days
This will allow for example printing of two landscape A4 pages that can
be put underneath each other.

We only support this for the full schedule, not room-specific one - the room
specific ones will take so little space anyway since they are not rendered to
a relative-to-per-second size.

Splitting of pages will *only* happen on cross-schedule sessions. For each
page split we will pick the closest one to the number of pages requested and
split there. The split session (e.g. "lunch") is included on both pages, to
make it easier to follow. (We could make that configurable in the future, perhaps?)