Add support for object-results-reports across years
Only for result reports, and only if an object is selected, allow
a report that spans years. This can get a summary view for projects
that cross years.

For any cross-year selection, the entire set of entries will always
be used, it cannot be limited by an end date (the entry box is removed).