Add support for sending an email upon registration completion
For each conference it's now possible to define a message text
that will be sent out to all attendees once their registration
is confirmed (either by invoice payment, bulk invoice payment,
no invoice needed, or manual processing). This should help with
the complaints from people who are used to getting this (the
status was already visible on the website), as well as give
a chance to include some extra information.

With this we can now replace the mailman integration with builtin
emails whenever needed.