Add ability to "turn off" a sponsorship level
The level will still show everywhere, but it won't be possible to sign
up for it.

Typically used for limited-number-of-sponsors levels, when they've run out.

Do this with a boolean instead of a "max count", both to cover cases where
contract is pending discussions (don't want an extra signup during that),
and to cover any other case than just the limited number case.