Re-enable CSRF framework
Seems this was only enabled in one of my feature branches. It should
definitely be enabled, to make sure it protects the forms of the site.

There may be some forms that are now broken - there might be more views
that require @csrf_exempt. We'll have to fix those one by one.