Don't crash when deleting an invoice for a confirmed sponsor
This can happen when a sponsor is manually confirmed, due to contract
signing, but invoice not paid yet.

Fix this by in general handling exceptions properly for invoice cancellations,
even though this is the only case where the problem can show up today. But
it could show up elsewhere in the future.