Send email when a sponsor is confirmed
When a sponsor was confirmed by direct invoice payment they would receive
the receipt, but a sponsor confirmed by signing a contract was actually
never sent notification of that unless the person clicking the button
remembered to do it manually.

Take this chance to also explain in the email a bit about how the benefits,
emailing and multiple managers work.