Add ability for sponsors to buy vouchers and discount codes
Make it possible for sponsors to buy vouchers and discount codes
directly on the sponsorship dashboard.

Vouchers will be paid immediately, and become active once paid,
just as before. The only difference is that the sponsor can place
the order directly on the site, instead of having to email the
staff who adds it manually.

Discount codes are requested, at which point the code immediately
becomes active. Once the code reaches it's end (either by reaching
the end date that it's valid for, or by reaching it's maximum
number of uses), an invoice will be generated for the sum of all
the discounts given out, to be paid by the sponsor.

This means discount code now track both sponsors, sponsor reps,
and invoices. They can still be added manually of course. If added
with a sponsor, they will automatically be invoiced. If added without
a sponsor, they will not be invoiced at all.

Usage of both vouchers and discount codes can now be viewed on
the sponsor dashboard, and this is also where new ones are ordered.

The vouchers that are claimed as a sponsorship benefit will show
up on the listing on the dashboard, but obviously not generate an invoice.

The ability to buy vouchers and discount codes are controlled
by flags on the sponsorship level - so it's possible to let different
levels sponsors have access to none or one of the two only.

A new cronjob has been added that should run daily in the morning,
and will generate the invoices for discount codes that are being
closed (either because they expired or because they hit their
maximum usage).