Add support for refunding sponsorship invoices
This wasn't considered as something we'd ever do, I guess, but the
Belfort conference clearly shows that we need the ability.

On unconfirmation, we don't actually unconfirm all the benefits -
that part is left up to the operator. But we flag the whole sponsorship
as unconfirmed, so that no further benefits can be claimed.