Add a new payment provider, Braintree
This is a US based creditcard payment provider. Primarily to be used
by pgopen at this point, but added in a generic way so it can be reused
should someone else want it in the future.

The user flow of payment is slightly different from the previous
providers. Instead of sending the user to the payment site itself,
which we do for Adyen and Paypal, Braintree have decided to use client side
javascript. So the flow is.

1. Request a nonce using the braintree server side APIs, and render it in a form
2. The Braintree javascript will rewrite the creditcard number etc into a
   secure token, that is submitted back to the site.
3. With this token, a transaction is generated server-side in the braintree
   APIs, and the user is redirected to the payment completed page.

The transaction goes the same basic flow as the Adyen ones, with slightly
different terminlogy: Authorized -> Settled -> Disbursed

The first step is handled by the views in braintreepayment/views.py, and
the second two steps are both handled by the new cronjob in
tools/braintreepayment/update_transaction_status.py which should run
regularly (probably every 6 hours or so).

Another new cronjob, tools/braintreepayment/log_reporter.py, must be run
in order to get alerts for anything abnormal happening in the system. Really
bad things are emailed immediately, but most things are just logged as
errors and sent as an aggregated email by this script.

Note that unlike the other payment providers, the braintree module has
an external dependency. It requires the "braintree" module installed
(from pip or packages). For this reason, the braintree module is
disabled by default, and has to be enabled by setting the value
ENABLE_BRAINTREE=True in local_settings.py.