Create "attendee dashboard" page
For now this basically just replaces the "your registration is complete"
screen, but the idea is we can put more things on here as they show up,
which really don't belong on the registration complete screen.

At this point, the only real "dashboardy" feature is the list of emails
sent out to attendees.