Add ability to post-order additional options
Also update the registration dashboard to show some interesting
details about the registration, instead of just a placeholder.

This makes it possible to order things like training even if the main
registration has completed, to help people who change their mind.

This will also automatically upgrade the reservation to one that is
required by the options, if necessary. If the upgraded registration
type is more expensive than the current one, automatically add the
cost difference to the invoice.

To control this, implement the option "upsellable" for each additional
option, without which an option cannot be added, and an option
"upsell_target" on a registration type, without which a registration
cannot be changed to that type.