Add support for automatically canceling invoices
Each invoice can be tagged with a timestamp when it will be automatically
canceled. If this time passes without payment, it will be by the new
cronjob cancel_invoices.py which is intended to run every hour or so
(depending on required granularity for invoice canceling).

If an invoice is flagged for autocancel, the timestamp for this will
be included in the reminder if sent. It will also be shown on the web
payment page for the invoice once the due date has expired.

By default, no invoices are flagged for automatic cancel. It's up to each
subsystem to set that manually, or for the operator who's adding a manual
invoice.