Implement invoice auto-cancel for conference registrations
Make it possible to specify an auto-cancel time (in hours) on
a conference, a registration type and/or an additional option.
If any of those is specified, the shortest time will be picked.
If none is specified, then of course the invoice will never be
autocancelled.

The idea is we can use this for example when a specific training is
about to sell out, and do faster cancels of those. Once we have
proper waitlist management implemented, it will also use this
system to ensure people don't "tie up" limited slots.