Add basic waitlist management
This allows a maximum number of attendees before waitlist to be
specified on the conference record. Once this number is reached,
people trying to register will automatically be given the option
to sign up on the waitlist instead. They can then be approved one
by one off the waitlist with a limited time offer to finish the
signup. If the time expires, they get moved back to the bottom of
the waitlist, and then we repeat the process.

The limit should normally be set a bit below the actual limit of
the venue, so that we can control who gets the last signups. For
example, we might need to reserve some for a speaker or a sponsor.