More clearly show state of registrations when building bulk registration
Instead of just always saying that an email doesn't exist, explicitly
tell the user when an existing *invoice* exists, as this is a typical
mistake (the end-user proceeding past the point that clearly says don't
proceed past this point if using bulk registration). Also indicate
if the registration is part of a different bulk registration (a case
never spotted in reality, but you never know).

If the registration is *completed*, however, we still respond with
a generic message, so it's not possible to use this as a backdoor to
enumerate who has registered for the conference.