Ensure registration dashboard is rendered even with registration closed
As we now use the registration dashboard for things like showing emails
that were sent to all attendees, it would be bad to just hide them when
we close registration.

We only show the dashboard if a registration is confirmed in this case,
and still render the "registration closed" message for any new attendees.