Implement basic conference wiki support
This is a limited wiki for conference attendees, intended for things
like ride sharing or restaurant coordination.

The wiki supports simple markdown syntax, and nothing more than that.

Only admins can create pages (through the dashboard). Each page can
be assigned permissions (public, based on regtype, or individual
attendee) to view and/or edit the pages. *Only* confirmed registrations
will ever be able to see anything on the wiki (unlike previous times
when we've used public pages on the main postgresql.org wiki).

If at least one page is visible to the attendee, the section shows
up on the registration dashboard.

Attendees can subscribe to changes to a page, in which case they will
receive an email when the page is edited. Changes are always sent to
the conference contact address (for post-moderation if required).

History is tracked on all pages, and it's possible to view the history
of a page including diff between versions (unless disabled on individual
wiki pages from the admin interface).

To implement the diff functionality, import the diffablemodel code from
the pgcommitfest app that does similar things, and build on top of that.