Automatically remove additional options from old unconfirmed registrations
Registrations that don't have invoices or bulk paymens still "lock" additional
options from others. And some attendees never complete their registrations
at all.

With this new cronjob enabled, those will also respect the "invoice autocancel hours"
parameter from the additional options, and be removed after that time. As there is
no actual registration, the only thing that happens is that the current registration
is edited and the additional option removed, the rest of it remains. And as long
as nobody else manages to register in between, the attendee can go back and re-register
for it as long as it's still available.

Expiry also happens automatically when an invoice is cancelled, as it makes more
sense to do that right away rather than waiting for the next cronjob to run, for
consistency.

This adds a "lastmodified" field to all conference registrations, which is used to
track this and other changes. So we only track the changes on the level of the
whole registration, and thus won't catch people who go in and edit their address
for example, without modifying the options. But the target with these removals
are those that have completely abandoned their registration, which will be
properly handled.