Implement basic support for Adyen Bank Transfer payments
This abstracts out the Adyen provider into two, one that does the
classic creditcard payments and one that does the bank transfer.

For the bank transfer one, enforce locale to France, otherwise the
option just comes up completely empty if Adyen thinks the IP address
is outside Europe. While it is of course possible to use a European
bank account even if you'e travelling outside the area.

Not enabled by default, and there is more work that needs to be done
before it can be, but this serves a starting point for testing.