Update Adyen integration to use SHA-256 signatures
Our current SHA1 signature method has been deprecated by Adyen, and
it turns out the newer one is easier to use anyway (as all fields are
now signed, meaning we don't need to keep track of which). As a bonus
we get the stronger security of SHA-256.

NOTE! This requires a new signing key to be updated in local_settings.py,
and this has to be done in sync with publishing of the Skin on the Adyen
website!