Include the long url format in invoice reminders by default
If possible, include the secret link to the invoice and not the one that
reqiures a login. People often forward the invoice to somebody else, and if
we include the secret link then the recipient of that forward (typically
someone in the finance department) will still be able to view the payment
information.