Split the waitlist view in two sections
The first section will be all pending entries on the list, the
second will hold those that have been given an offer and completed it.

This makes it possible to get to the "make new offer" button without
a huge amount of scrolling even if many offers off the waitlist have
been completed. The page still gets fairly large, but the scrolling
is removed.