Track displayname and url for sponsors
Tracking displayname makes it possible for the sponsor to have a different
name on the invoice than what's used on the website, and explicitly asking
for the URL means we don't have to email about that one every time.

These fields are tracked and stored, but not actually used for anything
internally.