Track which payment provider was used to pay an invoice
It's a "lossy tracking", since we allow invoices that are not paid with
an explicit payment method, to maintain some of the decoupling that existed.
With this, we can now also show for each invoice how much the payment fees
were, which will make it easier to deal with refunds.