Implement proper refund management
This means a few things:
1. We track refunds properly. Each invoice can be refunded once, but we now
   also support partial refunds (which is important, as almost every refund we
   make on a conference registration ends up being a partial one - if nothing
   else it is deducted transaction fees).
2. We support API initiated refunds. That means that as long as the payment is
   done using a supported provider (today that means Adyen and Paypal), the
   complete invoice processing is done in the webapp, which means there is no
   risk of getting the wrong numbers into the accounting (which has happened more
   than once).
3. We now generate proper refund notices in PDF format for all refunds (both
   automated and manual). The user receives this one as well as status updates
   throughout the refund process.

Actual API refunds are handled by a new cronjob, to ensure that we don't end up
blocking the user if the API is slow. Initiating the refund puts it on the queue,
the cronjob sends it to the provider, and the notification from the provider
flags it as completed (and generates the refund notice).