Partial reimplement of layout for conference admin pages
Instead of forcing them into the restricted framework of the main website,
set up a very plain bootstrap environment. This makes it a *lot* easier to
use on mobile devices as a start.

Also redo the dashboard page completely to focus on current and upcoming
conferences, and avoiding the "jumping" of conferences up the list that
often happened before. And properly fix the per-conference split that
was done before, by updating all the "back" links to not ruin it.

Most of the pages aren't actually converted to bootstrap, so they still
look a bit so-so, but they are a lot closer to being mobile-friendly
than they were before. More work to be done...