Support values to entries in signup form
This means we can have "will attend with a plus one" count as 2 already
in the signup, making it a lot easier to handle limitations on signups.
And also to make "No" count as zero...

If nothing is entered in the field for limits, it will work just like before
and just count the number of entries.