Fix more django 1.8isms
queryset was renamed go get_queryset in ModelAdmin. This only affected
things in /admin/. For Additional Options and for Prepaid Batches it
caused a crash, noted by Stephen Frost. For conference sessions and
registrations it would let somebody who is a full admin on one
conference view existing registrations on a different one. Oops.