Implement global http auth middleware
When configured this will require a fixed username/password to access
all pages on the site. The reason for this is to be able to more easily
deploy a testing system that's actually reachable outside localhost
without exposing it to other users and/or search engines.

A new decorator, global_login_exempt, is also created to exclude certain
URLs from this password requirement. Currently only used for Adyen
notifications, and it's main usecase is to support receiving pingbacks
from external systems such as payment providers.