Callforpapers form needs to be atomic
This is probably the reason for occassional session-without-speaker
creations. If for some reason adding the speaker to the session fails
(should be no real reason for this, but perhaps trying to add the same
speaker more than once or something like that could trigger it), the
session was left around but not attached to any speaker.